+++ src/main/java/io/supertokens/pluginInterface/thirdparty/exception/ThirdPartyException.java
@@ -1,0 +1,5 @@
+package io.supertokens.pluginInterface.thirdparty.exception;
+
+public class ThirdPartyException extends Exception {
+    private static final long serialVersionUID = -8903928332291556225L;
+}